Give pavucontrol options their own section
This moves the box containing the checkboxes to the top of the
Configuration tab and adds a label with an icon to act as a header of
sorts. Margins and paddings are set where relevant so that everything
lines up with the first section of other tabs when switching them.